wip move publishing methods to publishing context
