+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/workspace/document-workspace.context.ts
@@ -2,20 +2,15 @@ import { UmbDocumentTypeDetailRepository } from '../../document-types/repository
 import { UmbDocumentPropertyDatasetContext } from '../property-dataset-context/document-property-dataset.context.js';
 import type { UmbDocumentDetailRepository } from '../repository/index.js';
 import { UMB_DOCUMENT_DETAIL_REPOSITORY_ALIAS } from '../repository/index.js';
-import type { UmbDocumentVariantPublishModel, UmbDocumentDetailModel, UmbDocumentVariantModel } from '../types.js';
-import { UmbDocumentPublishingRepository } from '../publishing/index.js';
+import type { UmbDocumentDetailModel, UmbDocumentVariantModel } from '../types.js';
 import {
 	UMB_CREATE_DOCUMENT_WORKSPACE_PATH_PATTERN,
 	UMB_CREATE_FROM_BLUEPRINT_DOCUMENT_WORKSPACE_PATH_PATTERN,
 	UMB_DOCUMENT_COLLECTION_ALIAS,
 	UMB_DOCUMENT_ENTITY_TYPE,
-	UMB_DOCUMENT_PUBLISH_MODAL,
-	UMB_DOCUMENT_PUBLISH_WITH_DESCENDANTS_MODAL,
 	UMB_DOCUMENT_SAVE_MODAL,
-	UMB_DOCUMENT_SCHEDULE_MODAL,
 	UMB_EDIT_DOCUMENT_WORKSPACE_PATH_PATTERN,
 } from '../constants.js';
-import { UmbDocumentValidationRepository } from '../repository/validation/document-validation.repository.js';
 import { UmbDocumentPreviewRepository } from '../repository/preview/index.js';
 import { UMB_DOCUMENT_DETAIL_MODEL_VARIANT_SCAFFOLD, UMB_DOCUMENT_WORKSPACE_ALIAS } from './constants.js';
 import type { UmbEntityModel } from '@umbraco-cms/backoffice/entity';
@@ -26,15 +21,7 @@ import {
 	UmbWorkspaceIsNewRedirectControllerAlias,
 } from '@umbraco-cms/backoffice/workspace';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
-import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
-import {
-	UmbRequestReloadChildrenOfEntityEvent,
-	UmbRequestReloadStructureForEntityEvent,
-} from '@umbraco-cms/backoffice/entity-action';
-import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
-import { UmbServerModelValidatorContext } from '@umbraco-cms/backoffice/validation';
 import { UmbDocumentBlueprintDetailRepository } from '@umbraco-cms/backoffice/document-blueprint';
-import { UMB_NOTIFICATION_CONTEXT } from '@umbraco-cms/backoffice/notification';
 import {
 	UmbContentDetailWorkspaceContextBase,
 	type UmbContentCollectionWorkspaceContext,
@@ -43,7 +30,7 @@ import {
 import type { UmbDocumentTypeDetailModel } from '@umbraco-cms/backoffice/document-type';
 import { UmbIsTrashedEntityContext } from '@umbraco-cms/backoffice/recycle-bin';
 import { UMB_APP_CONTEXT } from '@umbraco-cms/backoffice/app';
-import UmbUnpublishDocumentEntityAction from '../publishing/unpublish/entity-action/unpublish.action.js';
+import { UmbDocumentPublishingRepository } from '../publishing/index.js';
 
 type ContentModel = UmbDocumentDetailModel;
 type ContentTypeModel = UmbDocumentTypeDetailModel;
@@ -60,10 +47,12 @@ export class UmbDocumentWorkspaceContext
 		UmbPublishableWorkspaceContext,
 		UmbContentCollectionWorkspaceContext<UmbDocumentTypeDetailModel>
 {
+	/**
+	 * The publishing repository for the document workspace.
+	 * @deprecated Will be removed in v17. Use the methods on the UMB_DOCUMENT_PUBLISHING_WORKSPACE_CONTEXT instead.
+	 * @memberof UmbDocumentWorkspaceContext
+	 */
 	public readonly publishingRepository = new UmbDocumentPublishingRepository(this);
-
-	#serverValidation = new UmbServerModelValidatorContext(this);
-	#validationRepository?: UmbDocumentValidationRepository;
 
 	readonly isTrashed = this._data.createObservablePartOfCurrent((data) => data?.isTrashed);
 	readonly contentTypeUnique = this._data.createObservablePartOfCurrent((data) => data?.documentType.unique);
@@ -219,6 +208,10 @@ export class UmbDocumentWorkspaceContext
 		return this._handleSubmit();
 	}
 
+	public async saveAndPreview(): Promise<void> {
+		return this.#handleSaveAndPreview();
+	}
+
 	async #handleSaveAndPreview() {
 		const unique = this.getUnique();
 		if (!unique) throw new Error('Unique is missing');
@@ -251,212 +244,43 @@ export class UmbDocumentWorkspaceContext
 		previewWindow?.focus();
 	}
 
-	async #handleSaveAndPublish() {
-		const unique = this.getUnique();
-		if (!unique) throw new Error('Unique is missing');
-
-		let variantIds: Array<UmbVariantId> = [];
-
-		const { options, selected } = await this._determineVariantOptions();
-
-		// If there is only one variant, we don't need to open the modal.
-		if (options.length === 0) {
-			throw new Error('No variants are available');
-		} else if (options.length === 1) {
-			// If only one option we will skip ahead and save the document with the only variant available:
-			variantIds.push(UmbVariantId.Create(options[0]));
-		} else {
-			// If there are multiple variants, we will open the modal to let the user pick which variants to publish.
-			const modalManagerContext = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
-			const result = await modalManagerContext
-				.open(this, UMB_DOCUMENT_PUBLISH_MODAL, {
-					data: {
-						options,
-						pickableFilter: this._readOnlyLanguageVariantsFilter,
-					},
-					value: { selection: selected },
-				})
-				.onSubmit()
-				.catch(() => undefined);
-
-			if (!result?.selection.length || !unique) return;
-
-			variantIds = result?.selection.map((x) => UmbVariantId.FromString(x)) ?? [];
-		}
-
-		const saveData = await this._data.constructData(variantIds);
-		await this._runMandatoryValidationForSaveData(saveData);
-
-		// Create the validation repository if it does not exist. (we first create this here when we need it) [NL]
-		this.#validationRepository ??= new UmbDocumentValidationRepository(this);
-
-		// We ask the server first to get a concatenated set of validation messages. So we see both front-end and back-end validation messages [NL]
-		if (this.getIsNew()) {
-			const parent = this.getParent();
-			if (!parent) throw new Error('Parent is not set');
-			this.#serverValidation.askServerForValidation(
-				saveData,
-				this.#validationRepository.validateCreate(saveData, parent.unique),
-			);
-		} else {
-			this.#serverValidation.askServerForValidation(
-				saveData,
-				this.#validationRepository.validateSave(saveData, variantIds),
-			);
-		}
-
-		// TODO: Only validate the specified selection.. [NL]
-		return this.validateAndSubmit(
-			async () => {
-				return this.#performSaveAndPublish(variantIds, saveData);
-			},
-			async () => {
-				// If data of the selection is not valid Then just save:
-				await this._performCreateOrUpdate(variantIds, saveData);
-				// Notifying that the save was successful, but we did not publish, which is what we want to symbolize here. [NL]
-				const notificationContext = await this.getContext(UMB_NOTIFICATION_CONTEXT);
-				// TODO: Get rid of the save notification.
-				// TODO: Translate this message [NL]
-				notificationContext.peek('danger', {
-					data: { message: 'Document was not published, but we saved it for you.' },
-				});
-				// Reject even thought the save was successful, but we did not publish, which is what we want to symbolize here. [NL]
-				return await Promise.reject();
-			},
-		);
-	}
-
-	async #performSaveAndPublish(variantIds: Array<UmbVariantId>, saveData: UmbDocumentDetailModel): Promise<void> {
-		const unique = this.getUnique();
-		if (!unique) throw new Error('Unique is missing');
-
-		await this._performCreateOrUpdate(variantIds, saveData);
-
-		const { error } = await this.publishingRepository.publish(
-			unique,
-			variantIds.map((variantId) => ({ variantId })),
-		);
-
-		if (!error) {
-			const eventContext = await this.getContext(UMB_ACTION_EVENT_CONTEXT);
-			const event = new UmbRequestReloadStructureForEntityEvent({
-				unique: this.getUnique()!,
-				entityType: this.getEntityType(),
-			});
-
-			eventContext.dispatchEvent(event);
-
-			this._closeModal();
-		}
-	}
-
-	public async publish() {
+	/**
+	 * @deprecated Will be removed in v17.
+	 */
+	async publish() {
 		throw new Error('Method not implemented.');
 	}
 
-	public async saveAndPreview(): Promise<void> {
-		return this.#handleSaveAndPreview();
+	/**
+	 * Save the document and publish it.
+	 * @deprecated Will be removed in v17. Use the UMB_DOCUMENT_PUBLISHING_WORKSPACE_CONTEXT instead.
+	 */
+	async saveAndPublish(): Promise<void> {
+		alert('WRONG METHOD');
 	}
 
-	public async saveAndPublish(): Promise<void> {
-		return this.#handleSaveAndPublish();
-	}
-
+	/**
+	 * Schedule the document to be published at a later date.
+	 * @deprecated Will be removed in v17. Use the UMB_DOCUMENT_PUBLISHING_WORKSPACE_CONTEXT instead.
+	 */
 	public async schedule() {
-		const { options, selected } = await this._determineVariantOptions();
-
-		const modalManagerContext = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
-		const result = await modalManagerContext
-			.open(this, UMB_DOCUMENT_SCHEDULE_MODAL, {
-				data: {
-					options,
-					pickableFilter: this._readOnlyLanguageVariantsFilter,
-				},
-				value: { selection: selected.map((unique) => ({ unique, schedule: {} })) },
-			})
-			.onSubmit()
-			.catch(() => undefined);
-
-		if (!result?.selection.length) return;
-
-		// Map to the correct format for the API (UmbDocumentVariantPublishModel)
-		const variants =
-			result?.selection.map<UmbDocumentVariantPublishModel>((x) => ({
-				variantId: UmbVariantId.FromString(x.unique),
-				schedule: x.schedule,
-			})) ?? [];
-
-		if (!variants.length) return;
-
-		// TODO: Validate content & Save changes for the selected variants — This was how it worked in v.13 [NL]
-
-		const unique = this.getUnique();
-		if (!unique) throw new Error('Unique is missing');
-		await this.publishingRepository.publish(unique, variants);
+		alert('WRONG METHOD');
 	}
 
+	/**
+	 * Unpublish the document.
+	 * @deprecated Will be removed in v17. Use the UMB_DOCUMENT_PUBLISHING_WORKSPACE_CONTEXT instead.
+	 */
 	public async unpublish() {
-		const unique = this.getUnique();
-		const entityType = this.getEntityType();
-		if (!unique) throw new Error('Unique is missing');
-		if (!entityType) throw new Error('Entity type is missing');
-
-		// TODO: remove meta
-		new UmbUnpublishDocumentEntityAction(this, { unique, entityType, meta: {} as never }).execute();
+		alert('WRONG METHOD');
 	}
 
+	/**
+	 * Publish the document and all its descendants.
+	 * @deprecated Will be removed in v17. Use the UMB_DOCUMENT_PUBLISHING_WORKSPACE_CONTEXT instead.
+	 */
 	public async publishWithDescendants() {
-		const unique = this.getUnique();
-		if (!unique) throw new Error('Unique is missing');
-
-		const entityType = this.getEntityType();
-		if (!entityType) throw new Error('Entity type is missing');
-
-		const { options, selected } = await this._determineVariantOptions();
-
-		const modalManagerContext = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
-		const result = await modalManagerContext
-			.open(this, UMB_DOCUMENT_PUBLISH_WITH_DESCENDANTS_MODAL, {
-				data: {
-					options,
-					pickableFilter: this._readOnlyLanguageVariantsFilter,
-				},
-				value: { selection: selected },
-			})
-			.onSubmit()
-			.catch(() => undefined);
-
-		if (!result?.selection.length) return;
-
-		// Map to variantIds
-		const variantIds = result?.selection.map((x) => UmbVariantId.FromString(x)) ?? [];
-
-		if (!variantIds.length) return;
-
-		const { error } = await this.publishingRepository.publishWithDescendants(
-			unique,
-			variantIds,
-			result.includeUnpublishedDescendants ?? false,
-		);
-
-		if (!error) {
-			const eventContext = await this.getContext(UMB_ACTION_EVENT_CONTEXT);
-
-			// request reload of this entity
-			const structureEvent = new UmbRequestReloadStructureForEntityEvent({
-				entityType,
-				unique,
-			});
-
-			// request reload of the children
-			const childrenEvent = new UmbRequestReloadChildrenOfEntityEvent({
-				entityType,
-				unique,
-			});
-
-			eventContext.dispatchEvent(structureEvent);
-			eventContext.dispatchEvent(childrenEvent);
-		}
+		alert('WRONG METHOD');
 	}
 
 	public createPropertyDatasetContext(
